Move EMFManipulationUtils to common project; refactor TGG code to use the utils class
